react-version/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq>=0.4.0
langchain>=0.1.0
langgraph>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.2.2
tiktoken>=0.5.2
flask-jwt-extended==4.6.0
bcrypt==4.1.2
cryptography==41.0.7

# Production server with async/event-driven support
gunicorn==21.2.0
gevent==23.9.1
